src/requirements.txt
aioredis==1.1.0
alog==0.9.13
asgiref==2.3.2
async-timeout==3.0.0
atomicwrites==1.2.0
attrs==18.1.0
autobahn==18.8.1
Automat==0.7.0
borax==1.1.2
certifi==2018.8.24
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.2
diff-match-patch==20121119
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1
django-appconf==1.0.2
django-bootstrap3==10.0.1
django-compressor==2.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-echarts==0.3.6
django-environ==0.4.5
django-filter==2.0.0
django-import-export==1.0.1
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.16.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
jupyter-echarts-pypkg==0.1.0
lml==0.0.2
MarkupSafe==1.0
more-itertools==4.3.0
msgpack==0.5.6
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.5
Pillow==5.2.0
pluggy==0.7.1
py==1.6.0
pyecharts==0.4.1
pyecharts-jupyter-installer==0.0.3
PyHamcrest==1.9.0
pytest==3.7.4
pytest-asyncio==0.9.0
pytest-django==3.4.2
pytest-pythonpath==0.7.3
python-dotenv==0.9.1
pytz==2018.5
PyYAML==3.13
rcssmin==1.0.6
requests==2.19.1
rjsmin==1.0.12
schedule==0.5.0
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
tablib==0.12.1
Twisted==18.7.0
txaio==18.7.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
websocket-client==0.51.0
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.5.0
